+++ src/sass/modal.scss
@@ -1,143 +1,59 @@
-@import url('https://fonts.googleapis.com/css?family=Montserrat');
-
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
 body {
-  font-family: 'Montserrat', sans-serif;
-  background-color: #eee;
-  color: #111;
-  height: 200vh;
-}
-button {
-  margin: 0.75rem;
-  padding: 0.25rem;
-  cursor: pointer;
+  font-family: Arial, sans-serif;
+  margin: 0;
 }
 
-/* Modal */
+.movie-container {
+  display: flex;
+  flex-wrap: wrap; /* Permite cardurilor să se așeze pe mai multe linii */
+  justify-content: space-around; /* Distribuie cardurile uniform pe orizontală */
+}
+
+.movie-card {
+  border: 1px solid #ddd;
+  margin: 10px;
+  padding: 10px;
+  cursor: pointer;
+  width: 200px; /* Lățimea dorită pentru card pe ecrane mari */
+}
+
+@media screen and (max-width: 600px) {
+  .movie-card {
+    width: 100%; /* Pe ecrane mici, lățimea cardului va fi 100% */
+  }
+}
+
 .modal {
+  display: none;
   position: fixed;
-  left: 0;
   top: 0;
+  left: 0;
   width: 100%;
   height: 100%;
-  display: none;
+  background-color: rgba(0, 0, 0, 0.5);
   justify-content: center;
-  align-items: flex-start; /* or center */
-  z-index: 1000; /* depending on what z-index values exist in your page, in order for the modal layer to sit on top of other content */
-  overflow: auto; /* enable scrolling if needed - although this is not really necessary since we are going take care of overflow in modal-content */
-  background-color: rgba(0, 0, 0, 0.8);
+  align-items: center;
 }
-.modal.modal-show {
-  animation: fadeIn 0.1s ease-in-out forwards;
-}
-.modal.modal-hide {
-  animation: fadeOut 0.1s ease-in-out 0.1s forwards;
-}
+
 .modal-content {
-  position: relative;
-  background-color: #fff;
-  margin: 2rem;
-  padding: 2rem;
-  border-radius: 0.25rem;
-  width: 50%;
-  max-height: 75%;
-  overflow: auto;
-  box-shadow: 0 4px 20px rgba(0, 0, 0, 0.4);
-}
-.modal.modal-show .modal-content {
-  animation: fadeInDown 0.3s ease-in-out forwards;
-}
-.modal.modal-hide .modal-content {
-  animation: fadeOutUp 0.2s ease-in-out forwards;
-}
-.modal-content h1 {
+  background-color: #fefefe;
+  padding: 20px;
+  border-radius: 8px;
   text-align: center;
-  margin-bottom: 2rem;
+  max-width: 600px; /* Lățimea maximă a conținutului modal pe ecrane mari */
+  width: 100%; /* Asigură că modalul ocupă întreaga lățime */
 }
-.modal-content p {
-  margin: 1rem 0;
-  line-height: 1.5rem;
-}
+
 .modal-close {
   position: absolute;
-  top: 0;
-  right: 0;
-  font-size: 1.75rem;
-  font-weight: bold;
-  padding: 0 0.75rem;
-  /* width: 2rem;
-  height: 2rem;
-  display: flex;
-  justify-content: center;
-  align-items: center; */
-  color: rgba(0, 0, 0, 0.2);
+  top: 10px;
+  right: 10px;
+  font-size: 20px;
   cursor: pointer;
-  user-select: none;
-}
-.modal-close:hover,
-.modal-close:focus {
-  color: rgba(0, 0, 0, 0.5);
 }
 
-/* Animations */
-@keyframes fadeIn {
-  0% {
-    opacity: 0;
-  }
-  100% {
-    opacity: 1;
-  }
-}
-@keyframes fadeOut {
-  0% {
-    opacity: 1;
-  }
-  100% {
-    opacity: 0;
-  }
-}
-@keyframes fadeInDown {
-  0% {
-    opacity: 0;
-    transform: translateY(-3rem);
-  }
-  100% {
-    opacity: 1;
-    transform: translateY(0);
-  }
-}
-@keyframes fadeOutUp {
-  0% {
-    opacity: 1;
-    transform: translateY(0);
-  }
-  100% {
-    opacity: 0;
-    transform: translateY(-3rem);
-  }
-}
-
-@media (max-width: 992px) {
-  html {
-    font-size: 14px;
-  }
+@media screen and (max-width: 600px) {
   .modal-content {
-    width: 80%;
-  }
-}
-@media (max-width: 767px) {
-  html {
-    font-size: 12px;
-  }
-  .modal-content {
-    padding: 2rem 1rem 1rem 1rem;
-    width: 90%;
-  }
-  .modal-content h1 {
-    margin-bottom: 1.5rem;
+    max-width: 80%; /* Pe ecrane mici, conținutul modal va avea lățimea maximă */
   }
 }
